work

+++ src/components/pages/MaterialUnitsPage.vue
@@ -1,19 +1,23 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue';
-import Button from 'primevue/button';
 import Dialog from 'primevue/dialog';
 import BaseModelCrudTable from '@/components/tables/BaseModelCrudTable.vue';
 import type { BaseModelCrudTableConfig } from '@/types/table';
 import { useMaterialsLibraryStore } from '../../stores/materialsLibrary';
 import { storeToRefs } from 'pinia';
 import MaterialUnit from '../../models/MaterialUnit';
-import { useConfirm } from 'primevue/useconfirm';
-import { useToast } from 'primevue/usetoast';
 import MaterialUnitForm from '@/components/forms/MaterialUnitForm.vue';
 import { cloneDeep } from 'lodash';
+import { useToastMessage } from '@/composables/useToastMessage';
+import { useConfirmMessage } from '@/composables/useConfirmMessage';
+import { useAuthApi } from '@/composables';
+import type BaseCollection from '@/models/BaseCollection';
 
-const confirm = useConfirm();
-const toast = useToast();
+const pageLoading = ref(true);
+
+const { addSuccessMessage, addErrorMessage } = useToastMessage();
+const { confirmDeleteMessage } = useConfirmMessage();
+
 const materialsLibraryStore = useMaterialsLibraryStore();
 
 const { units, unitsLoading } = storeToRefs(materialsLibraryStore);
@@ -30,8 +34,23 @@ const config = ref<BaseModelCrudTableConfig<MaterialUnit>>({
     startEdit: (model: MaterialUnit) => handleStartEdit(model),
     startDelete: (model: MaterialUnit) => handleDeleteUnit(model),
     columns: [
-        { field: 'name', header: 'Наименование', class: 'w-[50%]', filter: true },
-        { field: 'label', header: 'Короткое наименование', class: 'w-[25%]', filter: true },
+        { 
+            field: 'name', 
+            header: 'Наименование', 
+            class: 'w-[50%]', 
+            filter: true, 
+        },
+        { 
+            field: 'label', 
+            header: 'Короткое наименование', 
+            class: 'w-[25%]', 
+            filter: true, 
+        },
+        {
+            field: 'coefficient',
+            header: 'Коэффициент',
+            class: 'w-[25%]',
+        },
     ]
 });
 
@@ -48,60 +67,59 @@ const handleUpdateUnit = async (model: MaterialUnit) => {
     materialsLibraryStore.updateUnit(model)
         .then(() => {
             isEditDialogOpen.value = false;
-            toast.add({ severity: 'success', summary: 'Успешно', detail: 'Единица измерения успешно обновлена', life: 3000 });
-            // materialsLibraryStore.loadUnits();
+            addSuccessMessage('Единица измерения успешно обновлена');
         })
         .catch((error) => {
-            toast.add({ severity: 'error', summary: 'Ошибка', detail: 'Не удалось обновить единицу измерения', life: 3000 });
-            console.error('Error updating unit:', error);
+            addErrorMessage('Не удалось обновить единицу измерения');
+            console.error(error);
         });
 }
 
 const handleStoreUnit = async (model: MaterialUnit) => {
-    materialsLibraryStore.storeUnit(model)
+    await materialsLibraryStore.storeUnit(model)
         .then(() => {
             isAddDialogOpen.value = false;
-            toast.add({ severity: 'success', summary: 'Успешно', detail: 'Единица измерения успешно добавлена', life: 3000 });
-            // materialsLibraryStore.loadUnits();
+            addSuccessMessage('Единица измерения успешно добавлена');
         })
         .catch((error) => {
-            toast.add({ severity: 'error', summary: 'Ошибка', detail: 'Не удалось добавить единицу измерения', life: 3000 });
-            console.error('Error storing unit:', error);
+            addErrorMessage('Не удалось добавить единицу измерения');
+            console.error(error);
         });
 }
 
 const handleDeleteUnit = async (model: MaterialUnit) => {
-
-    confirm.require({
-        message: 'Вы уверены, что хотите удалить эту единицу измерения?',
-        header: 'Удаление единицы измерения',
+    confirmDeleteMessage({
         accept: async () => {
             materialsLibraryStore.deleteUnit(model)
                 .then(() => {
-                    toast.add({ severity: 'success', summary: 'Успешно', detail: 'Единица измерения успешно удалена', life: 3000 });
-                    materialsLibraryStore.loadUnits();
+                    addSuccessMessage('Единица измерения успешно удалена');
                 })
                 .catch((error) => {
-                    toast.add({ severity: 'error', summary: 'Ошибка', detail: 'Не удалось удалить единицу измерения', life: 3000 });
-                    console.error('Error deleting unit:', error);
+                    addErrorMessage('Не удалось удалить единицу измерения');
+                    console.error(error);
                 });
         },
     });
 }
 
 onMounted(() => {
-    materialsLibraryStore.loadUnits();
+    materialsLibraryStore.loadUnits()
+        .then(() => {
+            pageLoading.value = false;
+        });
 })
 </script>
 
 <template>
-    <BaseModelCrudTable :models="units.toArray()" :models-loading="unitsLoading" :config="config" />
+    <BaseModelCrudTable :models="units.toArray()" :models-loading="pageLoading" :config="config" />
 
     <Dialog v-model:visible="isEditDialogOpen" header="Редактирование единицы измерения" modal>
-        <MaterialUnitForm v-model:model="editableUnit" @submit="handleUpdateUnit" @cancel="isEditDialogOpen = false" />
+        <MaterialUnitForm v-model:model="editableUnit" :loading="unitsLoading" @submit="handleUpdateUnit"
+            @cancel="isEditDialogOpen = false" />
     </Dialog>
 
     <Dialog v-model:visible="isAddDialogOpen" header="Добавление единицы измерения" modal>
-        <MaterialUnitForm v-model:model="newUnit" @submit="handleStoreUnit" @cancel="isAddDialogOpen = false" />
+        <MaterialUnitForm v-model:model="newUnit" :loading="unitsLoading" @submit="handleStoreUnit"
+            @cancel="isAddDialogOpen = false" />
     </Dialog>
 </template>
